refactored get_function_benchmark_timings and get_benchmark_timings into BenchmarkDatabaseUtils class
